NEEDS TESTING: Presumed event race condition
I don't have an AR Drone, and I don't currently have enough room to adapt and test this with a bebop, so it bears verifying. The code looked like it was launching the drone prior to attaching the "while flying" code as an event triggered lambda, so I moved liftoff to occur after this event is attached.